requirements.txt
asgiref==3.2.5
async-timeout==3.0.1
autobahn==20.3.1
Automat==0.6.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
Click==7.0
colorama==0.3.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
daphne==2.4.1
decorator==4.3.0
django-rest-swagger==2.2.0
django-tables2==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.10.3
docutils==0.15.2
drf-yasg==1.17.0
entrypoints==0.3
fasteners==0.12.0
Flask==1.0.2
future==0.16.0
GeoIP==1.3.2
hiredis==1.0.1
html5lib==1.0.1
httplib2==0.11.3
hyperlink==17.3.1
idna==2.6
ifaddr==0.1.6
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.5.0
keyring==18.0.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
logilab-common==1.4.3
loremipsum==1.0.5
macaroonbakery==1.2.3
Mako==1.0.7
olefile==0.46
openapi-codec==1.3.2
packaging==19.1
pandocfilters==1.4.2
paramiko==2.6.0
pexpect==4.6.0
pickleshare==0.7.5
protobuf==3.6.1
pylint==2.2.2
pymacaroons==0.13.0
pymongo==3.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyRFC3339==1.1
python-dateutil==2.7.3
python-debian==0.1.36
python3-openid==3.1.0
pytz==2019.2
pyxattr==0.6.1
PyYAML==5.1.2
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
rdflib==4.2.2
redis==2.10.6
reportlab==3.5.23
requests==2.21.0
requests-unixsocket==0.1.5
roman==2.0.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scipy==1.2.2
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.16.0
